move AccountMeta to its own file
