+++ sdk/instruction/src/lib.rs
@@ -17,12 +17,14 @@
 use borsh::BorshSerialize;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::wasm_bindgen;
-use {crate::error::InstructionError, solana_pubkey::Pubkey, solana_sanitize::Sanitize};
 #[cfg(feature = "serde")]
 use {
     serde_derive::{Deserialize, Serialize},
     solana_short_vec as short_vec,
 };
+use {solana_pubkey::Pubkey, solana_sanitize::Sanitize};
+pub mod account_meta;
+pub use account_meta::AccountMeta;
 pub mod error;
 #[cfg(target_os = "solana")]
 pub mod syscalls;
@@ -271,104 +273,6 @@ impl Instruction {
             program_id,
             accounts,
             data: data.to_vec(),
-        }
-    }
-}
-
-/// Describes a single account read or written by a program during instruction
-/// execution.
-///
-/// When constructing an [`Instruction`], a list of all accounts that may be
-/// read or written during the execution of that instruction must be supplied.
-/// Any account that may be mutated by the program during execution, either its
-/// data or metadata such as held lamports, must be writable.
-///
-/// Note that because the Solana runtime schedules parallel transaction
-/// execution around which accounts are writable, care should be taken that only
-/// accounts which actually may be mutated are specified as writable. As the
-/// default [`AccountMeta::new`] constructor creates writable accounts, this is
-/// a minor hazard: use [`AccountMeta::new_readonly`] to specify that an account
-/// is not writable.
-#[repr(C)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-pub struct AccountMeta {
-    /// An account's public key.
-    pub pubkey: Pubkey,
-    /// True if an `Instruction` requires a `Transaction` signature matching `pubkey`.
-    pub is_signer: bool,
-    /// True if the account data or metadata may be mutated during program execution.
-    pub is_writable: bool,
-}
-
-impl AccountMeta {
-    /// Construct metadata for a writable account.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use solana_pubkey::Pubkey;
-    /// # use solana_instruction::{AccountMeta, Instruction};
-    /// # use borsh::{BorshSerialize, BorshDeserialize};
-    /// #
-    /// # #[derive(BorshSerialize, BorshDeserialize)]
-    /// # #[borsh(crate = "borsh")]
-    /// # pub struct MyInstruction;
-    /// #
-    /// # let instruction = MyInstruction;
-    /// # let from = Pubkey::new_unique();
-    /// # let to = Pubkey::new_unique();
-    /// # let program_id = Pubkey::new_unique();
-    /// let instr = Instruction::new_with_borsh(
-    ///     program_id,
-    ///     &instruction,
-    ///     vec![
-    ///         AccountMeta::new(from, true),
-    ///         AccountMeta::new(to, false),
-    ///     ],
-    /// );
-    /// ```
-    pub fn new(pubkey: Pubkey, is_signer: bool) -> Self {
-        Self {
-            pubkey,
-            is_signer,
-            is_writable: true,
-        }
-    }
-
-    /// Construct metadata for a read-only account.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use solana_pubkey::Pubkey;
-    /// # use solana_instruction::{AccountMeta, Instruction};
-    /// # use borsh::{BorshSerialize, BorshDeserialize};
-    /// #
-    /// # #[derive(BorshSerialize, BorshDeserialize)]
-    /// # #[borsh(crate = "borsh")]
-    /// # pub struct MyInstruction;
-    /// #
-    /// # let instruction = MyInstruction;
-    /// # let from = Pubkey::new_unique();
-    /// # let to = Pubkey::new_unique();
-    /// # let from_account_storage = Pubkey::new_unique();
-    /// # let program_id = Pubkey::new_unique();
-    /// let instr = Instruction::new_with_borsh(
-    ///     program_id,
-    ///     &instruction,
-    ///     vec![
-    ///         AccountMeta::new(from, true),
-    ///         AccountMeta::new(to, false),
-    ///         AccountMeta::new_readonly(from_account_storage, false),
-    ///     ],
-    /// );
-    /// ```
-    pub fn new_readonly(pubkey: Pubkey, is_signer: bool) -> Self {
-        Self {
-            pubkey,
-            is_signer,
-            is_writable: false,
         }
     }
 }
